requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
_tflow_select=2.3.0=mkl
absl-py=0.13.0=py36h06a4308_0
aiohttp=3.7.4=py36h27cfd23_1
astor=0.8.1=py36h06a4308_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py36h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py36h06a4308_0
boto3=1.17.97=pyhd3eb1b0_0
botocore=1.20.97=pyhd3eb1b0_1
brotlipy=0.7.0=py36h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.10.8=ha878542_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.5.30=py36h06a4308_0
cffi=1.14.5=py36h261ae71_0
chardet=3.0.4=py36h06a4308_1003
click=8.0.1=pyhd3eb1b0_0
coverage=5.5=py36h27cfd23_2
cryptography=3.4.7=py36hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py36_0
cython=0.29.23=py36h2531618_0
dataclasses=0.8=pyh4f3eec9_6
dbus=1.13.18=hb2f20db_0
decorator=5.0.9=pyhd3eb1b0_0
dgl-cuda10.2=0.6.1=py36_0
dgllife=0.2.6=py36_0
expat=2.4.1=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=py36h5fab9bb_3
gast=0.3.3=py_0
gensim=4.0.1=py36h2531618_0
glib=2.68.2=h36276a3_0
google-api-core=1.25.1=pyhd3eb1b0_0
google-auth=1.32.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-cloud-core=1.7.0=pyhd3eb1b0_0
google-cloud-storage=1.31.0=py_0
google-crc32c=1.1.2=py36h27cfd23_0
google-pasta=0.2.0=py_0
google-resumable-media=1.3.1=pyhd3eb1b0_1
googleapis-common-protos=1.53.0=py36h06a4308_0
googledrivedownloader=0.4=pyhd3deb0d_1
grpcio=1.36.1=py36h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py36hd6299e0_1
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
idna_ssl=1.1.0=py36h06a4308_0
importlib-metadata=3.10.0=py36h06a4308_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py36h5ca1d4c_0
ipython=7.16.1=py36h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.17.0=py36_0
jinja2=3.0.2=pyhd8ed1ab_0
jmespath=0.10.0=py_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py36h06a4308_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py36h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=8_mkl
libcrc32c=1.1.1=he6710b0_2
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1d223b6_11
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
llvm-openmp=12.0.1=h4bd325d_1
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py36h06a4308_0
markupsafe=2.0.1=py36h8f6f2f9_0
matplotlib=3.3.4=py36h06a4308_0
matplotlib-base=3.3.4=py36h62a2d02_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py36he8ac12f_0
mkl_fft=1.3.0=py36h54f3939_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=5.1.0=py36h27cfd23_2
ncurses=6.2=he6710b0_1
networkx=2.5=py_0
ninja=1.10.2=h4bd325d_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py36_0
openssl=1.1.1l=h7f98852_0
opt_einsum=3.3.0=pyhd3eb1b0_1
pandas=1.1.5=py36ha9443f7_0
parso=0.8.2=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickle5=0.0.11=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py36he98fc37_0
pip=21.1.2=py36h06a4308_0
plotly=4.14.3=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.15.8=py36hc4f0c31_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyg=2.0.1=py36_torch_1.8.0_cu101
pygments=2.9.0=pyhd3eb1b0_0
pyjwt=1.7.1=py36_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h06a4308_0
python=3.6.13=h12debd9_1
python-dateutil=2.8.1=pyhd3eb1b0_0
python-louvain=0.15=pyhd3deb0d_0
python_abi=3.6=2_cp36m
pytorch=1.8.0=cpu_py36h2d15a6b_1
pytorch-cluster=1.5.9=py36_torch_1.8.0_cu101
pytorch-scatter=2.0.8=py36_torch_1.8.0_cu101
pytorch-sparse=0.6.12=py36_torch_1.8.0_cu101
pytorch-spline-conv=1.2.1=py36_torch_1.8.0_cu101
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py36h8f6f2f9_1
pyzmq=20.0.0=py36h2531618_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
retrying=1.3.3=py36_2
rsa=4.7.2=pyhd3eb1b0_1
s3transfer=0.4.2=pyhd3eb1b0_0
scikit-learn=0.24.2=py36ha9443f7_0
scipy=1.5.2=py36h0b6359f_0
setuptools=52.0.0=py36h06a4308_0
sip=4.19.8=py36hf484d3e_0
six=1.16.0=pyhd3eb1b0_0
sleef=3.5.1=h9b69904_2
smart_open=5.1.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
stellargraph=1.2.1=py_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.2.0=mkl_py36h5a57954_0
tensorflow-base=2.2.0=mkl_py36hd506778_0
tensorflow-estimator=2.5.0=pyh7b7c402_0
termcolor=1.1.0=py36h06a4308_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.1=py36h27cfd23_0
tqdm=4.61.1=pyhd3eb1b0_1
traitlets=4.3.3=py36_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py36h7b6447c_1
xz=5.2.5=h7b6447c_0
yacs=0.1.6=py_0
yaml=0.2.5=h516909a_0
yarl=1.6.3=py36h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
